refactor: Provider 제거
